+++ src/lib/components/ui/Navbar.svelte
@@ -123,6 +123,16 @@
             <Icon mini src={GlobeAlt} size="16" slot="icon" />
             <span class="hidden md:inline">Explore</span>
         </Button>
+
+        <!---Modlog--->
+        <Button
+            href="/modlog"
+            aria-label="Modlog"
+            class="max-md:w-9 max-md:h-8 max-md:!p-0 dark:text-zinc-300 text-slate-700 hover:text-inherit hover:dark:text-inherit hover:bg-slate-200 hover:border-slate-300"
+        >
+            <Icon mini src={Newspaper} size="16" slot="icon" />
+            <span class="hidden md:inline">Modlog</span>
+        </Button>
       
         <!--- 'Create' Menu--->
         <Menu alignment="bottom-right">